paymentsummary.js renamed to paymentsummary.jsx and data added to display amount charged on the payment summary page
